requirements.txt
pyspark==4.0.0
delta-spark==4.0.0
boto3==1.34.0
pytest==8.2.0
